test(runtime): isolate ENTER CRITICAL YIELD policy regression
